Update versions of pandas, sqlalchemy, numpy, tensorflow, and scikit-learn. (#84)
* 📌 Updated versions for pandas, tensorflow, scikit-learn, sqlalchemy, numpy,

* 📌 Updated versions for pandas, tensorflow, scikit-learn, sqlalchemy, numpy,

* 🔖 Removed support for Python 3.9, added to CHANGELOG.md

* 🔖 Added support to Python 3.11

* 🚨 Fixed linter issues. 🐛 Fixed bug checking utc timezone.

* 🚨 Fixed more linting issues

* 🚨 Fixed black compatibility

* 🔖 Removed support for python 3.11

* 🐛 onehot features in feature engineer now return int32

* 🐛 Added numeric_only instead of skipna in the rolling median.

* 🐛 Changed keras argument from "lr" to "learning_rate"

* 🚨 Fixed linting issue

* 📝 Fixed docstring example to have the right History object.

* 🚨 Fixed warnings, and made some examples working.

* 🚨 Fixed some whitespace

* 🚨 Fixed some whitespace

* 🚨 Linting?

* 🐛 Fixes the SHAP example.

* 🎨➖ Removes dependency on eli5, and now creates own feature importance scores.

* 🎨 changed pandas hour character in some files

* 🚨 Applied linter

* 📝 Added documentation and removed ELI5 comments in code.

---------

+++ sam/feature_engineering/simple_feature_engineering.py
@@ -178,7 +178,7 @@ class SimpleFeatureEngineer(BaseFeatureEngineer):
                 for value in range(comp_min, comp_max + 1):
                     comp_series = self._get_time_column(X, component)
                     colname_ = f"{colname}_{value}"
-                    X_out[colname_] = (comp_series == value).astype(int)
+                    X_out[colname_] = (comp_series == value).astype("int32")
 
             elif type == "cyclical":
                 comp_series = self._get_time_column(X, component)
